Rename Block to Podcast Player
It has been decided by Dave Martin that Podcast Player is a better name.
