--- extensions/blocks/podcast-episodes/editor.scss
@@ -1,5 +1,0 @@
-/**
- * Editor styles for Podcast Episodes
- */
-
-.wp-block-jetpack-podcast-episodes { }
