+++ extensions/blocks/podcast-player/editor.scss
@@ -1,0 +1,5 @@
+/**
+ * Editor styles for Podcast Player
+ */
+
+.wp-block-jetpack-podcast-player { }
