+++ extensions/index.json
@@ -28,5 +28,5 @@
 		"videopress",
 		"wordads"
 	],
-	"beta": [ "amazon", "seo", "podcast-episodes" ]
+	"beta": [ "amazon", "seo", "podcast-player" ]
 }
